Fix bold text followed by underlined text
The value of last was not updated upon encountering a star, so it would
contain whatever character happened to be before the star. If that
happened to be an alphanumeric character,
parse_format(p, FORMAT_UNDERLINE); was never called since it was
detected as a case of an underscore between alphanumeric characters.